some collision improvements

+++ src/main/java/de/ellpeck/rockbottom/api/entity/player/AbstractEntityPlayer.java
@@ -34,9 +34,9 @@ import java.util.List;
 
 public abstract class AbstractEntityPlayer extends EntityLiving implements ICommandSender{
 
-    public static final float MOVE_SPEED = 0.2F;
-    public static final float CLIMB_SPEED = 0.15F;
-    public static final float RANGE = 5;
+    public static final double MOVE_SPEED = 0.2;
+    public static final double CLIMB_SPEED = 0.15;
+    public static final double RANGE = 5;
 
     public AbstractEntityPlayer(IWorld world){
         super(world);
